split up commands by type into traits that can be composed
