+++ modules/core/src/main/scala/com/github/scoquelin/arugula/commands/internal/LettuceRedisServerAsyncCommands.scala
@@ -1,0 +1,23 @@
+package com.github.scoquelin.arugula.commands.internal
+
+import scala.concurrent.Future
+
+import com.github.scoquelin.arugula.api.commands.RedisServerAsyncCommands
+
+private[commands] trait LettuceRedisServerAsyncCommands[K, V] extends RedisServerAsyncCommands[K, V] with LettuceRedisCommandDelegation[K, V] {
+
+  override def flushAll: Future[Unit] =
+    delegateRedisClusterCommandAndLift(_.flushall()).map(_ => ())
+
+  override def info: Future[Map[String, String]] = {
+    def parseInfo(info: String): Map[String, String] =
+      info
+        .split("\\r?\\n")
+        .toList
+        .map(_.split(":", 2).toList)
+        .collect { case k :: v :: Nil => (k, v) }
+        .toMap
+
+    delegateRedisClusterCommandAndLift(_.info()).map(parseInfo)
+  }
+}
